requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.1
    # via streamlit
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
exceptiongroup==1.2.2
    # via pytest
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.23.0
    # via altair
numpy==2.2.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   plotly
    #   pytest
    #   streamlit
pandas==2.2.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==11.1.0
    # via streamlit
plotly==5.24.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
protobuf==5.29.3
    # via streamlit
pyarrow==19.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pytest==7.4.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2024.2
    # via pandas
referencing==0.36.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   streamlit
    #   tmdbsimple
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.41.1
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   plotly
    #   streamlit
tmdbsimple==2.9.1
    # via -r requirements.in
toml==0.10.2
    # via streamlit
tomli==2.2.1
    # via pytest
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   referencing
    #   rich
    #   streamlit
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
